Refactor PackagesPageComponent
